+++ samples/msal-angular-v2-samples/angular10-browser-sample/src/app/app.component.ts
@@ -34,7 +34,7 @@ export class AppComponent implements OnInit, OnDestroy {
         takeUntil(this._destroying$)
       )
       .subscribe((result) => {
-        console.log("APP COMPONENT - BROWSER BROADCAST: ", result);
+        console.log("EVENT: ", result);
         this.checkAccount();
       });
   }
