+++ samples/msal-angular-v2-samples/angular10-browser-sample/src/app/msal/msal.broadcast.service.ts
@@ -1,16 +1,21 @@
-import { Injectable } from '@angular/core';
-import { BroadcastService, BroadcastMessage, BroadcastEvent } from '@azure/msal-browser';
-
-// Broadcast service from msal-browser
+import { Inject, Injectable } from '@angular/core';
+import { BroadcastService, BroadcastEvent, IPublicClientApplication, AuthenticationResult, AuthError } from '@azure/msal-browser';
+import { Observable } from 'rxjs';
+import { MSAL_INSTANCE } from "./constants";
+  
 @Injectable()
-export class BrowserBroadcastService extends BroadcastService {
+export class BrowserBroadcastService {
+  // private broadcastService: BroadcastService;
 
+  constructor(
+    @Inject(MSAL_INSTANCE) private msalInstance: IPublicClientApplication,
+  ) {}
+
+  get msalSubject$(): Observable<any> {
+    return this.msalInstance.broadcastService.msalSubject$;
+  }
+
+  broadcast(type: BroadcastEvent, payload: AuthenticationResult | AuthError) {
+    this.msalInstance.broadcast(type, payload);
+  }
 }
-
-// export interface BrowserBroadcastMessage extends BroadcastMessage {
-
-// }
-
-// export enum BrowserBroadcastEvent {
-//   BroadcastEvent = BroadcastEvent,
-// }
