+++ samples/msal-angular-v2-samples/angular10-browser-sample/src/app/msal/msal.service.ts
@@ -38,31 +38,22 @@ export class MsalService implements IMsalService {
     acquireTokenPopup(request: AuthorizationUrlRequest): Observable<AuthenticationResult> {
         return from(
             this.msalInstance.acquireTokenPopup(request)
-                .then((authResponse) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
-                    return authResponse;
-                })
-                .catch((error: AuthError) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_FAILURE, error);
-                    throw error;
-                })
+                // .then((authResponse) => {
+                //     this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
+                //     return authResponse;
+                // })
         );
     }
     acquireTokenRedirect(request: RedirectRequest): Observable<void> {
-        this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_START, null);
         return from(this.msalInstance.acquireTokenRedirect(request));
     }
     acquireTokenSilent(silentRequest: SilentRequest): Observable<AuthenticationResult> {
         return from(
             this.msalInstance.acquireTokenSilent(silentRequest)
-                .then((authResponse: AuthenticationResult) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
-                    return authResponse;
-                })
-                .catch((error: AuthError) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_FAILURE, error);
-                    throw error;
-                })
+                // .then((authResponse: AuthenticationResult) => {
+                //     this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
+                //     return authResponse;
+                // })
         );
     }
     getAccountByUsername(userName: string): AccountInfo {
@@ -72,43 +63,38 @@ export class MsalService implements IMsalService {
         return this.msalInstance.getAllAccounts();
     }
     handleRedirectObservable(): Observable<AuthenticationResult> {
-        this.browserBroadcastService.broadcast(BroadcastEvent.HANDLE_REDIRECT_START, null);
         const loggedInAccounts = this.msalInstance.getAllAccounts();
         return from(
             this.msalInstance.handleRedirectPromise()
-                .then((authResponse: AuthenticationResult) => {
-                    if (authResponse) {
-                        const loggedInAccount = loggedInAccounts.find((account) => account.username === authResponse.account.username);
-                        if (loggedInAccount) {
-                            this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
-                        } else {
-                            this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_SUCCESS, authResponse);
-                        }
-                    }
-                    return authResponse;
-                })
-                .catch((error: AuthError) => {
-                    if (this.getAllAccounts().length > 0) {
-                        this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_FAILURE, error);
-                    } else {
-                        this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_FAILURE, error);
-                    }
-                    throw error;
-                })
+                // .then((authResponse: AuthenticationResult) => {
+                //     if (authResponse) {
+                //         const loggedInAccount = loggedInAccounts.find((account) => account.username === authResponse.account.username);
+                //         if (loggedInAccount) {
+                //             this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_SUCCESS, authResponse);
+                //         } else {
+                //             this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_SUCCESS, authResponse);
+                //         }
+                //     }
+                //     return authResponse;
+                // })
+                // .catch((error: AuthError) => {
+                //     if (this.getAllAccounts().length > 0) {
+                //         this.browserBroadcastService.broadcast(BroadcastEvent.ACQUIRE_TOKEN_FAILURE, error);
+                //     } else {
+                //         this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_FAILURE, error);
+                //     }
+                //     throw error;
+                // })
 
         );
     }
     loginPopup(request?: AuthorizationUrlRequest): Observable<AuthenticationResult> {
         return from(
             this.msalInstance.loginPopup(request)
-                .then((authResponse: AuthenticationResult) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_SUCCESS, authResponse);
-                    return authResponse;
-                })
-                .catch((error: AuthError) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_FAILURE, error);
-                    throw error;
-                })
+                // .then((authResponse: AuthenticationResult) => {
+                //     this.browserBroadcastService.broadcast(BroadcastEvent.LOGIN_SUCCESS, authResponse);
+                //     return authResponse;
+                // })
         );
     }
     loginRedirect(request?: RedirectRequest): Observable<void> {
@@ -120,14 +106,10 @@ export class MsalService implements IMsalService {
     ssoSilent(request: AuthorizationUrlRequest): Observable<AuthenticationResult> {
         return from(
             this.msalInstance.ssoSilent(request)
-                .then((authResponse: AuthenticationResult) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.SSO_SILENT_SUCCESS, authResponse);
-                    return authResponse;
-                })
-                .catch((error: AuthError) => {
-                    this.browserBroadcastService.broadcast(BroadcastEvent.SSO_SILENT_FAILURE, error);
-                    throw error;
-                })
+                // .then((authResponse: AuthenticationResult) => {
+                //     this.browserBroadcastService.broadcast(BroadcastEvent.SSO_SILENT_SUCCESS, authResponse);
+                //     return authResponse;
+                // })
         );
     }
